Improve CLI doc generator error handling
Instead of silently exiting with no error if it is not possible to read
the CLI doc generator config file, print an error message.

If there is no CLI generator config file, skip manual environment
variable additions and print a message. (Not using structured logging
since this will only ever be run manually and in CI.)